+++ aura-core/src/main/java/io/aura/AuraStarter.java
@@ -3,4 +3,7 @@ package io.aura;
 public interface AuraStarter {
     void start(Aura app);
     void stop();
+    default void reloadRoutes(Aura app) {
+        throw new UnsupportedOperationException("Hot reload not supported by this starter");
+    }
 }