+++ aura-db/src/main/java/io/aura/db/ThrowingRunnable.java
@@ -1,0 +1,6 @@
+package io.aura.db;
+
+@FunctionalInterface
+public interface ThrowingRunnable {
+    void run() throws Exception;
+}