+++ aura-db/src/main/java/io/aura/db/ThrowingSupplier.java
@@ -1,0 +1,6 @@
+package io.aura.db;
+
+@FunctionalInterface
+public interface ThrowingSupplier<T> {
+    T get() throws Exception;
+}